feat: use notmuch library to discover config and filter path
Previously, notcoal would use hardcoded paths to the notmuch
configuration and the filter rules file, alongside notmuch hooks.
However, notmuch offers number of fallback paths for the configuration
file and hooks directory. In some scenarios, notcoal could not find
notmuch configuration automatically, and users would need to specify the
path to the configuration file manually.

With this patch, notcoal will use the notmuch library bindings to open
a notmuch database. Any interaction with the notmuch configuration then
goes through the database. With this approach, notcoal discovers the
notmuch database and configuration file following the same logic as
notmuch itself. Similarly, the path to the filter rules file is based on
the path to notmuch's hooks directory.

Fixes #7.